Code Review Changes for Rent Minder Project.
